Fix dangling commas in a scripture
When parsing the verses, if there is a dangling comma because the
scripture is in a sentence, the last empty string in the "list" of
verses will be NaN and make it an invalid scripture.

Check for NaN, and just pass over it when parsing the verses.

For example:
"Read 1 Corinthians 2:16, and consider this question..."

It will think there are 2 verses and the second verse is NaN. By
skipping over this, we avoid this scenario.